Add --only option.
- Commands can no be executed selectively based on the extension(s) or
  tag(s) given.

+++ code/xopen_common.h
@@ -20,7 +20,7 @@ struct Instruction
 
 	// NOTE: Only one tag for now.
 	char *tag;
-	int tagLength;
+	size_t tagLength;
 };
 
 #endif
